Move functions needed to build project into setup.py
