+++ pypi_packaging.py
@@ -1,29 +1,7 @@
-import codecs
-
-import pkg_resources
 import os
+import pkg_resources
 
-
-def read(rel_path):
-    here = os.path.abspath(os.path.dirname(__file__))
-    with codecs.open(os.path.join(here, rel_path), "r") as fp:
-        return fp.read()
-
-
-def get_version(rel_path):
-    for line in read(rel_path).splitlines():
-        if line.startswith("__version__"):
-            delim = '"' if '"' in line else "'"
-            return line.split(delim)[1]
-    else:
-        raise RuntimeError("Unable to find version string.")
-
-
-# FORMAT: 1.x.x
-pypi_config = {
-    "version_target": get_version("cloudevents/__init__.py"),
-    "package_name": "cloudevents",
-}
+from setup import pypi_config
 
 
 def createTag():
